Refactor move createMutableHash() method from ScalarType to BeanProperty / BeanPropertyJsonMapper
